feat: stubbed out deletion highlighting
